refactor: ArchivePageService 인터페이스 분리한다
