+++ src/test/java/org/gsh/genidxpage/service/ArchivePageServiceTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class ArchivePageServiceTest {
+class AgileStoryArchivePageServiceTest {
 
     @DisplayName("아키이빙된 페이지 정보를 찾지 못했을 때 db에 기록한다")
     @Test
@@ -32,7 +32,7 @@ class ArchivePageServiceTest {
         );
         ApiCallReporter reporter = mock(ApiCallReporter.class);
 
-        ArchivePageService service = new ArchivePageService(caller, reporter);
+        AgileStoryArchivePageService service = new AgileStoryArchivePageService(caller, reporter);
 
         Assertions.assertThrows(ArchivedPageNotFoundExceptioin.class,
             () -> service.findArchivedPageInfo(dto));
@@ -57,7 +57,7 @@ class ArchivePageServiceTest {
         when(caller.isArchived(any())).thenReturn(true);
         ApiCallReporter reporter = mock(ApiCallReporter.class);
 
-        ArchivePageService service = new ArchivePageService(caller, reporter);
+        AgileStoryArchivePageService service = new AgileStoryArchivePageService(caller, reporter);
 
         service.findArchivedPageInfo(dto);
